refactor: separate glyphs component
